+++ src/HTML.hs
@@ -2,8 +2,12 @@ module HTML where
 
 import Control.Exception ( SomeException, catch )
 import Control.Monad ( guard )
+import Data.ByteString.Lazy ( toStrict )
 import Data.List ( isPrefixOf )
 import Network.HTTP.Conduit
+import Data.Text as T ( drop, dropEnd, pack, strip, unpack )
+import Data.Text.Encoding ( decodeUtf8 )
+import Text.HTML.TagSoup
 import Text.Regex.Posix ( (=~) )
 
 htmlTitle :: FilePath -> String -> IO (Maybe String)
@@ -11,10 +15,10 @@ htmlTitle regPath url = do
     regexps <- flip catch handleException . fmap lines $ readFile regPath 
     print regexps
     if (safeHost regexps url) then do
-      title <- flip catch handleException $ fmap (extractTitle . show) $ simpleHttp httpPrefixedURL
+      title <- flip catch handleException $ fmap (extractTitle . concat . lines . unpack . decodeUtf8 . toStrict) $ simpleHttp httpPrefixedURL
       case title of
         Just _ -> pure title
-        Nothing -> flip catch handleException $ fmap (extractTitle . show) $ simpleHttp httpsPrefixedURL
+        Nothing -> flip catch handleException $ fmap (extractTitle . unpack . decodeUtf8 . toStrict) $ simpleHttp httpsPrefixedURL
       else
         pure Nothing
   where
@@ -26,16 +30,19 @@ htmlTitle regPath url = do
 extractTitle :: String -> Maybe String
 extractTitle body = do
     guard (not $ null titleHTML)
-    pure $ chomp $ dropAround (length "<title>") (length "</title>") titleHTML
+    pure . escape . chomp $ removeMarker titleHTML
   where
     titleHTML :: String
     titleHTML = body =~ "<title>[^<]*</title>"
 
-dropAround :: Int -> Int -> String -> String
-dropAround s e = reverse . drop e . reverse . drop s
+removeMarker ::String -> String
+removeMarker = unpack . T.drop (length "<title>") . dropEnd (length "</title>") . pack
 
 chomp :: String -> String
-chomp = unwords . words . concat . lines
+chomp = unpack . strip . pack
+
+escape :: String -> String
+escape = fromTagText . head . parseTags
 
 -- Filter an URL so that we don’t make overviews of unknown hosts. Pretty
 -- cool to prevent people from going onto sensitive websites.